+++ src/components/app/App.tsx
@@ -1,10 +1,10 @@
 import { Routes, Route, Navigate } from 'react-router-dom';
 
-import Header from '../header/Header';
-import MainPage from '../mainPage/MainPage';
-import AboutPage from '../aboutPages/AboutPages';
-import FormPage from '../formPage/FormPage';
-import ErrorPage from '../errorPage/ErrorPage';
+import Layout from '../../pages/Layout/Layout';
+import MainPage from '../../pages/MainPage/MainPage';
+import AboutPages from '../../pages/AboutPage/AboutPage';
+import FormPage from '../../pages/FormPage/FormPage';
+import ErrorPage from '../../pages/NotFoundPage/ErrorPage';
 
 import './app.scss';
 
@@ -12,9 +12,9 @@ const App = () => {
   return (
     <div className="app">
       <Routes>
-        <Route path="/" element={<Header />}>
+        <Route path="/" element={<Layout />}>
           <Route index element={<MainPage />} />
-          <Route path="about" element={<AboutPage />} />
+          <Route path="about" element={<AboutPages />} />
           <Route path="form" element={<FormPage />} />
         </Route>
         <Route path="404" element={<ErrorPage />} />
