+++ src/components/header/Header.tsx
@@ -1,4 +1,4 @@
-import { Link, Outlet } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 import Nav from '../nav/Nav';
 
@@ -6,21 +6,18 @@ import './header.scss';
 
 const Header = () => {
   return (
-    <>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <h1>
-              <Link to="/" className="nav__link">
-                Logo
-              </Link>
-            </h1>
-            <Nav />
-          </div>
+    <header className="header">
+      <div className="container">
+        <div className="header__wrapper">
+          <h1>
+            <Link to="/" className="nav__link">
+              Logo
+            </Link>
+          </h1>
+          <Nav />
         </div>
-      </header>
-      <Outlet />
-    </>
+      </div>
+    </header>
   );
 };
