+++ src/pages/Layout/Layout.tsx
@@ -1,0 +1,13 @@
+import { Outlet } from 'react-router-dom';
+import Header from '../../components/header/Header';
+
+const Layout = () => {
+  return (
+    <>
+      <Header />
+      <Outlet />
+    </>
+  );
+};
+
+export default Layout;
