+++ src/pages/MainPage/mainPage.scss
